refactor: ContestPage안의 로직을 커스텀 훅으로 분리
